4 ontology test files into one
